added drawer to logsPage, created logsfile folder, added logs.tsx and css file to logsfile folder
