+++ ui/package.json
@@ -8,6 +8,7 @@
     "@docker/extension-api-client": "0.3.4",
     "@emotion/react": "11.10.4",
     "@emotion/styled": "11.10.4",
+    "@mui/icons-material": "^5.11.16",
     "@mui/material": "5.10.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
