+++ ui/src/pages/Logfiles/Logstyles.css
@@ -1,0 +1,13 @@
+.parentDiv{
+  display: flex;
+  
+}
+
+.leftdiv {
+  flex: 0 0 200px; 
+}
+
+
+.right-div {
+  flex: 0 0 200px; 
+}
